// studio/sanity.types.ts
export type LocationSection = {
  _type: 'locationSection'
  subtitle?: string
  title?: Array<{
    children?: Array<{
      marks?: Array<string>
      text?: string
      _type: 'span'
      _key: string
    }>
    style?: 'normal' | 'h2' | 'h3'
    listItem?: never
    markDefs?: Array<
      {
        _key: string
      } & TextColor
    >
    level?: number
    _type: 'block'
    _key: string
  }>
  text?: Array<{
    children?: Array<{
      marks?: Array<string>
      text?: string
      _type: 'span'
      _key: string
    }>
    style?: 'normal' | 'h2' | 'h3' | 'h4' | 'h5' | 'body2' | 'blockquote'
    listItem?: never
    markDefs?: Array<
      {
        _key: string
      } & TextColor
    >
    level?: number
    _type: 'block'
    _key: string
  }>
  topPadding?: boolean
  textAlign?: 'flex-start' | 'center'
}

export type ImageCarouselSection = {
  _type: 'imageCarouselSection'
  topPadding?: boolean
  images?: Array<{
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
    _key: string
  }>
}

export type TestimonialSection = {
  _type: 'testimonialSection'
  topPadding?: boolean
  backgroundColor?: SimplerColor
  testimonialTiles?: Array<{
    _ref: string
    _type: 'reference'
    _weak?: boolean
    _key: string
    [internalGroqTypeReferenceTo]?: 'quote'
  }>
}
// ...
export type Overlay = '0.1' | '0.2' | '0.3' | '0.4' | '0.5'

export type TextAlign = 'left' | 'center'
// ...
export type BlogSection = {
  _type: 'blogSection'
  subtitle?: string
  title?: Array<{
    children?: Array<{
      marks?: Array<string>
      text?: string
      _type: 'span'
      _key: string
    }>
    style?: 'normal' | 'h1' | 'h2' | 'h3'
    listItem?: never
    markDefs?: Array<{
      href?: string
      _type: 'link'
      _key: string
    }>
    level?: number
    _type: 'block'
    _key: string
  }>
  leftText?: Array<{
    children?: Array<{
      marks?: Array<string>
      text?: string
      _type: 'span'
      _key: string
    }>
    style?: 'normal' | 'body2' | 'h3' | 'h4' | 'h5' | 'blockquote'
    listItem?: 'bullet' | 'number'
    markDefs?: Array<{
      href?: string
      _type: 'link'
      _key: string
    }>
    level?: number
    _type: 'block'
    _key: string
  }>
  rightText?: Array<{
    children?: Array<{
      marks?: Array<string>
      text?: string
      _type: 'span'
      _key: string
    }>
    style?: 'normal' | 'h3' | 'h4' | 'h5' | 'blockquote'
    listItem?: 'bullet' | 'number'
    markDefs?: Array<{
      href?: string
      _type: 'link'
      _key: string
    }>
    level?: number
    _type: 'block'
    _key: string
  }>
  showArchive?: ShowArchive
  topPadding?: boolean
}
// ...
export type CtaSection = {
  _type: 'ctaSection'
  image?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
  title?: string
  text?: string
  links?: Array<
    {
      _key: string
    } & LinkDefined
  >
  leftAlign?: TextAlign
  topPadding?: boolean
  overlay?: Overlay
}

export type FeaturesListSection = {
  _type: 'featuresListSection'
  title?: Array<{
    children?: Array<{
      marks?: Array<string>
      text?: string
      _type: 'span'
      _key: string
    }>
    style?: 'normal' | 'h2' | 'h3'
    listItem?: never
    markDefs?: Array<
      {
        _key: string
      } & TextColor
    >
    level?: number
    _type: 'block'
    _key: string
  }>
  text?: Array<{
    children?: Array<{
      marks?: Array<string>
      text?: string
      _type: 'span'
      _key: string
    }>
    style?: 'normal' | 'h2' | 'h3' | 'h4' | 'h5' | 'body2' | 'blockquote'
    listItem?: never
    markDefs?: Array<
      {
        _key: string
      } & TextColor
    >
    level?: number
    _type: 'block'
    _key: string
  }>
  image?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
  topPadding?: boolean
  featuresTile?: Array<
    {
      _key: string
    } & FeaturesTile
  >
}

export type VideoSection = {
  _type: 'videoSection'
  topPadding?: boolean
  video?: VideoId
}

export type VideoId = {
  _type: 'videoId'
  url?: string
  id?: string
  service?: string
  thumbnail?: string
}

export type NewsletterSection = {
  _type: 'newsletterSection'
  image?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
  title?: Array<{
    children?: Array<{
      marks?: Array<string>
      text?: string
      _type: 'span'
      _key: string
    }>
    style?: 'normal' | 'h2' | 'h3' | 'h4' | 'h5' | 'body2' | 'blockquote'
    listItem?: never
    markDefs?: Array<
      {
        _key: string
      } & TextColor
    >
    level?: number
    _type: 'block'
    _key: string
  }>
  text?: Array<{
    children?: Array<{
      marks?: Array<string>
      text?: string
      _type: 'span'
      _key: string
    }>
    style?: 'normal' | 'h3' | 'h4' | 'h5' | 'body2' | 'blockquote'
    listItem?: 'bullet' | 'number'
    markDefs?: Array<{
      href?: string
      _type: 'link'
      _key: string
    }>
    level?: number
    _type: 'block'
    _key: string
  }>
  topPadding?: boolean
  mirror?: boolean
  backgroundColor?: SimplerColor
}

export type RecipesSection = {
  _type: 'recipesSection'
  backgroundColour?: SimplerColor
  showRecipesArchive?: ShowRecipesArchive
  verticalSpace?: VerticalSpace
}

export type TeamSection = {
  _type: 'teamSection'
  topPadding?: boolean
  subtitle?: string
  title?: Array<{
    children?: Array<{
      marks?: Array<string>
      text?: string
      _type: 'span'
      _key: string
    }>
    style?: 'normal' | 'h1' | 'h2' | 'h3'
    listItem?: never
    markDefs?: Array<{
      href?: string
      _type: 'link'
      _key: string
    }>
    level?: number
    _type: 'block'
    _key: string
  }>
  leftText?: Array<{
    children?: Array<{
      marks?: Array<string>
      text?: string
      _type: 'span'
      _key: string
    }>
    style?: 'normal' | 'body2' | 'h3' | 'h4' | 'h5' | 'blockquote'
    listItem?: 'bullet' | 'number'
    markDefs?: Array<{
      href?: string
      _type: 'link'
      _key: string
    }>
    level?: number
    _type: 'block'
    _key: string
  }>
  rightText?: Array<
    | {
        children?: Array<{
          marks?: Array<string>
          text?: string
          _type: 'span'
          _key: string
        }>
        style?: 'normal' | 'h3' | 'h4' | 'h5' | 'blockquote'
        listItem?: 'bullet' | 'number'
        markDefs?: Array<{
          href?: string
          _type: 'link'
          _key: string
        }>
        level?: number
        _type: 'block'
        _key: string
      }
    | {
        asset?: {
          _ref: string
          _type: 'reference'
          _weak?: boolean
          [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
        }
        hotspot?: SanityImageHotspot
        crop?: SanityImageCrop
        _type: 'image'
        _key: string
      }
  >
  links?: Array<
    {
      _key: string
    } & LinkDefined
  >
  tileColor?: SimplerColor
  teamTiles?: Array<{
    _ref: string
    _type: 'reference'
    _weak?: boolean
    _key: string
    [internalGroqTypeReferenceTo]?: 'teamMember'
  }>
}

export type HeaderSection = {
  _type: 'headerSection'
  image?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
  title?: string
  alignment?: TextAlign
  text?: Array<{
    children?: Array<{
      marks?: Array<string>
      text?: string
      _type: 'span'
      _key: string
    }>
    style?: 'normal' | 'body2'
    listItem?: never
    markDefs?: Array<
      {
        _key: string
      } & TextColor
    >
    level?: number
    _type: 'block'
    _key: string
  }>
  links?: Array<
    {
      _key: string
    } & LinkDefined
  >
  backgroundColour?: SimplerColor
  verticalSpace?: VerticalSpace
}
// ...
export type IngredientsList = Array<
  {
    _key: string
  } & IngredientItem
>
// ...
export type Recipes = {
  _id: string
  _type: 'recipes'
  _createdAt: string
  _updatedAt: string
  _rev: string
  metaTitle?: string
  metaDescription?: string
  ogTitle?: string
  ogDescription?: string
  ogImage?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
  locale?: 'en-US' | 'en-GB' | 'fr-FR'
  featuredMedia?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
  title?: string
  slug?: Slug
  modified?: string
  status?:
    | 'publish'
    | 'future'
    | 'draft'
    | 'pending'
    | 'private'
    | 'trash'
    | 'auto-draft'
    | 'inherit'
  duration?: Duration
  serves?: Serves
  date?: string
  ingredientsList?: IngredientsList
  author?: {
    _ref: string
    _type: 'reference'
    _weak?: boolean
    [internalGroqTypeReferenceTo]?: 'author'
  }
  instructions?: Array<
    | {
        children?: Array<{
          marks?: Array<string>
          text?: string
          _type: 'span'
          _key: string
        }>
        style?: 'normal' | 'body2' | 'h3' | 'h4' | 'h5' | 'blockquote'
        listItem?: 'bullet' | 'number'
        markDefs?: Array<
          {
            _key: string
          } & TextColor
        >
        level?: number
        _type: 'block'
        _key: string
      }
    | {
        asset?: {
          _ref: string
          _type: 'reference'
          _weak?: boolean
          [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
        }
        hotspot?: SanityImageHotspot
        crop?: SanityImageCrop
        _type: 'image'
        _key: string
      }
  >
  pageBuilder?: Array<
    | ({
        _key: string
      } & HeaderSection)
    | ({
        _key: string
      } & TestimonialSection)
    | ({
        _key: string
      } & TeamSection)
    | ({
        _key: string
      } & RecipesSection)
    | ({
        _key: string
      } & VideoSection)
    | ({
        _key: string
      } & FeaturesListSection)
    | ({
        _key: string
      } & CtaSection)
    | ({
        _key: string
      } & NewsletterSection)
    | ({
        _key: string
      } & ImageCarouselSection)
    | ({
        _key: string
      } & BlogSection)
    | ({
        _key: string
      } & StepsSection)
    | ({
        _key: string
      } & TimelineSection)
    | ({
        _key: string
      } & LocationSection)
    | ({
        _key: string
      } & ContactSection)
    | ({
        _key: string
      } & ClientLoginSection)
  >
  category?: {
    _ref: string
    _type: 'reference'
    _weak?: boolean
    [internalGroqTypeReferenceTo]?: 'recipesCategory'
  }
}
// ...
export type Serves = {
  _type: 'serves'
  serves?: number
  note?: string
}

export type Duration = {
  _type: 'duration'
  hours?: number
  minutes?: number
}
// ...
export type SanityImageCrop = {
  _type: 'sanity.imageCrop'
  top?: number
  bottom?: number
  left?: number
  right?: number
}

export type SanityImageHotspot = {
  _type: 'sanity.imageHotspot'
  x?: number
  y?: number
  height?: number
  width?: number
}
// ...
export type TextColor = {
  _type: 'textColor'
  label?: string
  value?: string
}
// ...
export type Slug = {
  _type: 'slug'
  current?: string
  source?: string
}
// ...
export declare const internalGroqTypeReferenceTo: unique symbol